mgr/cephadm: move HostAssignment to new module
(speeds up my editor)

(cherry picked from commit 524efe899426b73dc8c665b226a89dfd8c7ae05b)
